+++ files/index.d.ts
@@ -2614,26 +2614,6 @@ export function none<T>(predicate: (x: T) => boolean, list: T[]): boolean;
 export function none<T>(predicate: (x: T) => boolean): (list: T[]) => boolean;
 
 /*
-Method: nop
-
-Explanation: It returns `undefined`.
-
-Example:
-
-```
-const result = R.nop()
-// => undefined
-```
-
-Categories:
-
-Notes:
-
-*/
-// @SINGLE_MARKER
-export function nop(): void;
-
-/*
 Method: not
 
 Explanation: It returns a boolean negated version of `input`.
@@ -7490,6 +7470,24 @@ Notes:
 */
 // @SINGLE_MARKER
 export function omitPaths<T>(x: T): T;
+
+/*
+Method: noop
+
+Explanation:
+
+Example:
+
+```
+```
+
+Categories:
+
+Notes:
+
+*/
+// @SINGLE_MARKER
+export function noop(): void;
 
 // RAMBDAX_MARKER_END
 // ============================================