+++ source/nop-spec.ts
@@ -1,8 +1,0 @@
-import {nop} from 'rambda'
-
-describe('R.nop', () => {
-  it('call', () => {
-    const result = nop()
-    result // $ExpectType void
-  })
-})