+++ source/nop.spec.js
@@ -1,5 +1,0 @@
-import { nop } from './nop.js'
-
-test('call', () => {
-  expect(nop).not.toThrow()
-})