+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Synchronize with Ramda `0.29.0` release:
 
-- change order of `R.propEq` - [https://github.com/ramda/ramda/pull/2938/files](https://github.com/ramda/ramda/pull/2938/files)
+- change order of `R.propEq` - [Ramda MR](https://github.com/ramda/ramda/pull/2938/files)
 
 - `R.head`/`R.last` don't return `undefined` for non-empty arrays
 
@@ -10,6 +10,8 @@ Synchronize with Ramda `0.29.0` release:
 > it already did support dates in JS
 
 - Improve typings of `R.endsWith/startsWith` with regard to `string` input. - [PR #622](https://github.com/selfrefactor/rambda/pull/622)
+
+- Handle list as falsy value in `R.reduce` - [Ramda MR](https://github.com/ramda/ramda/pull/2997/files)
 
 7.5.0